requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
pydantic-settings==2.4.0
sqlalchemy==2.0.34
psycopg[binary]==3.2.2
python-multipart==0.0.9
PyMuPDF==1.24.10
python-docx==1.1.2
docxtpl==0.17.0
pgvector==0.3.2
numpy==2.0.1
tenacity==9.0.0
requests==2.32.3
httpx==0.27.0
fastembed==0.4.2
faster-whisper==1.1.1
surya-ocr          # layout-aware OCR replacing PaddleOCR (open source, USA origin)
pytesseract        # Tesseract Python binding — used as img2table OCR backend
img2table          # table detection from images/PDFs (uses Tesseract for cell text)
Pillow==10.4.0
convertdate==2.4.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
# Wake word detection — "Start Clerk" / "Over Clerk" via WebSocket
openwakeword==0.6.0
onnxruntime==1.19.2     # inference backend for openwakeword models
websockets==13.1        # used by uvicorn[standard] already; explicit pin for clarity
